refactor(workflow): extract validation and template helpers from executor.go (#140)
