+++ internal/workflow/executor.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
 
 	"github.com/giantswarm/muster/internal/api"
 	"github.com/giantswarm/muster/internal/template"
@@ -602,329 +601,4 @@ func (we *WorkflowExecutor) buildStepsArray(stepMetadata []stepMetadata, results
 	}
 
 	return steps
-}
-
-// executionContext holds the state during workflow execution
-type executionContext struct {
-	input        map[string]interface{} // Original input arguments
-	variables    map[string]interface{} // User-defined variables
-	results      map[string]interface{} // Results from previous steps
-	templateVars []string               // Track template variables used
-	stepMetadata []stepMetadata         // Track step metadata
-}
-
-// validateInputs validates the input arguments against the args definition
-func (we *WorkflowExecutor) validateInputs(argsDefinition map[string]api.ArgDefinition, args map[string]interface{}) error {
-	logging.Debug("WorkflowExecutor", "validateInputs called with args: %+v", args)
-	logging.Debug("WorkflowExecutor", "validateInputs args definition: %+v", argsDefinition)
-
-	// Check required fields and apply defaults
-	for key, argDef := range argsDefinition {
-		value, exists := args[key]
-
-		if !exists {
-			if argDef.Required {
-				logging.Error("WorkflowExecutor", fmt.Errorf("missing required field"), "Required field '%s' is missing from args %+v", key, args)
-				return fmt.Errorf("required field '%s' is missing", key)
-			}
-			// Apply default value if not provided
-			if argDef.Default != nil {
-				logging.Debug("WorkflowExecutor", "Applying default value for %s: %v", key, argDef.Default)
-				args[key] = argDef.Default
-			}
-			continue
-		}
-
-		// Basic type validation for provided values
-		if !we.validateType(value, argDef.Type) {
-			return fmt.Errorf("field '%s' has wrong type, expected %s", key, argDef.Type)
-		}
-	}
-
-	// Check for unknown arguments - allow extra properties for flexibility
-	// This follows the same pattern as ServiceClass.ValidateServiceArgs
-
-	logging.Debug("WorkflowExecutor", "validateInputs final args: %+v", args)
-	return nil
-}
-
-// validateType performs basic type validation
-func (we *WorkflowExecutor) validateType(value interface{}, expectedType string) bool {
-	switch expectedType {
-	case "string":
-		_, ok := value.(string)
-		return ok
-	case "number":
-		switch value.(type) {
-		case int, int32, int64, float32, float64:
-			return true
-		default:
-			return false
-		}
-	case "boolean":
-		_, ok := value.(bool)
-		return ok
-	case "array":
-		switch value.(type) {
-		case []interface{}, []string, []int, []float64:
-			return true
-		default:
-			return false
-		}
-	case "object":
-		_, ok := value.(map[string]interface{})
-		return ok
-	default:
-		return true // Unknown types pass validation
-	}
-}
-
-// resolveArguments resolves template variables in step arguments
-func (we *WorkflowExecutor) resolveArguments(args map[string]interface{}, ctx *executionContext) (map[string]interface{}, error) {
-	resolved := make(map[string]interface{})
-
-	for key, value := range args {
-		resolvedValue, err := we.resolveValue(value, ctx)
-		if err != nil {
-			return nil, fmt.Errorf("failed to render arguments for argument '%s': %w", key, err)
-		}
-		resolved[key] = resolvedValue
-	}
-
-	return resolved, nil
-}
-
-// resolveValue recursively resolves template variables in a value
-func (we *WorkflowExecutor) resolveValue(value interface{}, ctx *executionContext) (interface{}, error) {
-	switch v := value.(type) {
-	case string:
-		// Check if it contains a template pattern
-		if strings.Contains(v, "{{") && strings.Contains(v, "}}") {
-			return we.resolveTemplate(v, ctx)
-		}
-		return v, nil
-
-	case map[string]interface{}:
-		// Recursively resolve map values
-		resolved := make(map[string]interface{})
-		for k, val := range v {
-			resolvedVal, err := we.resolveValue(val, ctx)
-			if err != nil {
-				return nil, err
-			}
-			resolved[k] = resolvedVal
-		}
-		return resolved, nil
-
-	case []interface{}:
-		// Recursively resolve array elements
-		resolved := make([]interface{}, len(v))
-		for i, val := range v {
-			resolvedVal, err := we.resolveValue(val, ctx)
-			if err != nil {
-				return nil, err
-			}
-			resolved[i] = resolvedVal
-		}
-		return resolved, nil
-
-	default:
-		// Return other types as-is
-		return value, nil
-	}
-}
-
-// resolveTemplate resolves a template string
-func (we *WorkflowExecutor) resolveTemplate(templateStr string, ctx *executionContext) (interface{}, error) {
-	logging.Debug("WorkflowExecutor", "Resolving template: %s", templateStr)
-	logging.Debug("WorkflowExecutor", "Original results: %v", ctx.results)
-
-	// Track template variables used (extract from template string)
-	if strings.Contains(templateStr, ".input.") {
-		// Find all .input.variable_name patterns
-		words := strings.Fields(templateStr)
-		for _, word := range words {
-			if strings.Contains(word, ".input.") {
-				// Extract variable names like "input.service_name", "input.message"
-				if start := strings.Index(word, ".input."); start != -1 {
-					remaining := word[start+1:] // Remove the leading dot
-					if end := strings.IndexAny(remaining, " }"); end != -1 {
-						varName := remaining[:end]
-						if varName != "" && !contains(ctx.templateVars, varName) {
-							ctx.templateVars = append(ctx.templateVars, varName)
-						}
-					} else {
-						// Take everything until end of string, removing trailing }}
-						varName := strings.TrimSuffix(remaining, "}}")
-						varName = strings.TrimSuffix(varName, "}")
-						if varName != "" && !contains(ctx.templateVars, varName) {
-							ctx.templateVars = append(ctx.templateVars, varName)
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Create template context with original objects (no preprocessing)
-	templateCtx := map[string]interface{}{
-		"input":   ctx.input,
-		"vars":    ctx.variables,
-		"results": ctx.results,
-		"context": ctx.results, // Alias for results to support .context.variable syntax
-	}
-
-	logging.Debug("WorkflowExecutor", "Template context results (raw): %v", templateCtx["results"])
-
-	// Check if this is a simple variable access pattern ({{ .input.key }} or {{ .results.key }})
-	// If so, try to preserve the original type
-	if we.isSimpleVariableAccess(templateStr) {
-		originalValue := we.getOriginalValue(templateStr, ctx)
-		if originalValue != nil {
-			// For simple variable access, return the original value with its type preserved
-			return originalValue, nil
-		}
-	}
-
-	// Parse and execute template with strict mode options
-	result, err := we.template.RenderGoTemplate(templateStr, templateCtx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to render arguments: %w", err)
-	}
-
-	logging.Debug("WorkflowExecutor", "Template result: %v", result)
-	return result, nil
-}
-
-// isSimpleVariableAccess checks if the template is a simple variable access pattern
-func (we *WorkflowExecutor) isSimpleVariableAccess(templateStr string) bool {
-	// Remove whitespace and check if it matches {{ .input.key }} or {{ .results.key }} pattern
-	trimmed := strings.TrimSpace(templateStr)
-	if !strings.HasPrefix(trimmed, "{{") || !strings.HasSuffix(trimmed, "}}") {
-		return false
-	}
-
-	// Extract the inner part
-	inner := strings.TrimSpace(trimmed[2 : len(trimmed)-2])
-
-	// Check if it's a simple dot notation access
-	return strings.HasPrefix(inner, ".input.") || strings.HasPrefix(inner, ".results.") || strings.HasPrefix(inner, ".vars.")
-}
-
-// getOriginalValue extracts the original value from the context based on the template path
-func (we *WorkflowExecutor) getOriginalValue(templateStr string, ctx *executionContext) interface{} {
-	// Remove whitespace and extract the path
-	trimmed := strings.TrimSpace(templateStr)
-	if !strings.HasPrefix(trimmed, "{{") || !strings.HasSuffix(trimmed, "}}") {
-		return nil
-	}
-
-	// Extract the inner part
-	inner := strings.TrimSpace(trimmed[2 : len(trimmed)-2])
-
-	// Parse the path and get the value
-	if strings.HasPrefix(inner, ".input.") {
-		key := inner[7:] // Remove ".input."
-		return ctx.input[key]
-	} else if strings.HasPrefix(inner, ".results.") {
-		key := inner[9:] // Remove ".results."
-		return ctx.results[key]
-	} else if strings.HasPrefix(inner, ".vars.") {
-		key := inner[6:] // Remove ".vars."
-		return ctx.variables[key]
-	}
-
-	return nil
-}
-
-// contains checks if a string slice contains a specific string
-func contains(slice []string, item string) bool {
-	for _, s := range slice {
-		if s == item {
-			return true
-		}
-	}
-	return false
-}
-
-// validateJsonPath validates JSON path expectations against tool result
-func (we *WorkflowExecutor) validateJsonPath(toolResult *mcp.CallToolResult, jsonPathExpectations map[string]interface{}, execCtx *executionContext) (bool, error) {
-	// Parse the tool result as JSON
-	var resultData interface{}
-	if len(toolResult.Content) == 0 {
-		return false, fmt.Errorf("tool result has no content")
-	}
-
-	// Get the first content item and try to parse as JSON
-	if textContent, ok := toolResult.Content[0].(mcp.TextContent); ok {
-		if err := json.Unmarshal([]byte(textContent.Text), &resultData); err != nil {
-			return false, fmt.Errorf("failed to parse tool result as JSON: %w", err)
-		}
-	} else {
-		// If it's not text content, try to use it directly
-		resultData = toolResult.Content[0]
-	}
-
-	// Validate each JSON path expectation
-	for jsonPath, expectedValue := range jsonPathExpectations {
-		// Resolve the expected value template if it contains templating
-		resolvedExpectedValue, err := we.resolveValue(expectedValue, execCtx)
-		if err != nil {
-			return false, fmt.Errorf("failed to resolve expected value template for path %s: %w", jsonPath, err)
-		}
-
-		// Get the actual value from the result using simple path navigation
-		actualValue, err := we.getValueFromPath(resultData, jsonPath)
-		if err != nil {
-			logging.Debug("WorkflowExecutor", "Failed to get value from path %s: %v", jsonPath, err)
-			return false, nil // Path not found = condition fails
-		}
-
-		// Compare the values
-		if !we.valuesEqual(actualValue, resolvedExpectedValue) {
-			logging.Debug("WorkflowExecutor", "JSON path validation failed: path=%s, expected=%v, actual=%v",
-				jsonPath, resolvedExpectedValue, actualValue)
-			return false, nil
-		}
-
-		logging.Debug("WorkflowExecutor", "JSON path validation passed: path=%s, expected=%v, actual=%v",
-			jsonPath, resolvedExpectedValue, actualValue)
-	}
-
-	return true, nil
-}
-
-// getValueFromPath extracts a value from nested data using a simple path (e.g., "name", "data.field")
-func (we *WorkflowExecutor) getValueFromPath(data interface{}, path string) (interface{}, error) {
-	parts := strings.Split(path, ".")
-	current := data
-
-	for _, part := range parts {
-		switch v := current.(type) {
-		case map[string]interface{}:
-			if value, exists := v[part]; exists {
-				current = value
-			} else {
-				return nil, fmt.Errorf("path '%s' not found", path)
-			}
-		default:
-			return nil, fmt.Errorf("cannot navigate path '%s': not an object", path)
-		}
-	}
-
-	return current, nil
-}
-
-// valuesEqual compares two values for equality, handling type conversions
-func (we *WorkflowExecutor) valuesEqual(actual, expected interface{}) bool {
-	// Direct equality check first
-	if actual == expected {
-		return true
-	}
-
-	// Convert both to strings for comparison if they're different types
-	actualStr := fmt.Sprintf("%v", actual)
-	expectedStr := fmt.Sprintf("%v", expected)
-
-	return actualStr == expectedStr
 }
